+++ examples/kv_cache_reuse/remote_backends/resp/README.md
@@ -4,7 +4,7 @@ The cpp binding is in `csrc/redis/*`.
 
 The key optimization is multi-threading and batching on the C layer and the python side being awoken through the eventfd API where a callback consumes the completion belonging to a non-blocking submission. 
 
-An example of how to use the pybinding is in the sibling file `examples/kv_cache_reuse/remote_backends/resp/example_resp_client.py`.
+An GET / SET benchmark is in `examples/kv_cache_reuse/remote_backends/resp/example_resp_client.py`. The python client lives inside of `lmcache/v1/storage_backend/resp_client.py`.
 
 Install lmcache from source, then run a sanity check:
